backend/requirements.txt
fastapi
uvicorn
psycopg-pool
paho-mqtt
numpy
pandas
requests
python-dotenv